+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ###### 2023-08-14
 
 - Auto-Logout after 5 minutes of inactivity
+| subham maity |
+
+---
+
+###### 2020-08-14
+
+- Lead Generator page and backend controllers
+| Arian |
 
 ---